Use a dedicated token struct with position
I will need the position to report the location of the error in the original file in the later stages of processing. I had a choice between using a list and a struct. I went with a struct because it already fit with what I was going with, and created the setf macros for me.

I've been thinking of including the name of the file being processed as another field, but that's not really the information I need at the time of throwing an error, at least in the way the program is currently laid out.

I also had a choice between putting the file-position code in read-token (and then passing the position as an argument to the other read functions) and putting it directly in the read functions. I went with the latter for no other reason than that I already wrote it this way. I'm not sure what the correct choice is, but if a clear winner emerges in the future, I will switch.