Add conditions for unterminated $ and {
For now these are just basic conditions with nothing much going on,
and no restarts. I can imagine the format string used in :report would
make Larry Wall give a satisfied nod, so for the record here is a
little breakdown:

"~:[~:[~;~:*[?:~A] ~]~;~:*[~A:~:[?~;~:*~A~]] ~]Sexpcode..."
01. ~:[           start an if else, consuming the first argument
02. ~:[           if the first argument was nil, start another if else consuming the second argument
03. ~;            if second argument was nil, move to Sexpcode... without printing anything      [END]
04. ~:*           if second argument was not nil, back up to process the second argument again
05. [?:~A] ~]     print [?:position]<blank> and move to Sexpcode...                              [END]
06. ~:*           if first argument was not nil, back up to process the first argument again
07. [~A:          print [filename:
08. ~:[           start an if else, consuming second argument
09. ?~;           if second argument was nil, print ?
10. ~:*           if second argument was not nil, back up to process the second argument again
11. ~A~]] ~]      if second argument was not nil, print position]<blank> and move to Sexpcode... [END]

As always, 18. A Few FORMAT Recipes in PCL to the rescue. Or maybe to
make you spend 2 hours writing the format string and explanation for
it when you could have achieved the same via lisp code in 5
minutes. Although that would require repeating the format string, so I
guess this is fine.