Add function for preprocessing files
Added a not-so-well designed function PREPROCESS to test preprocessing
on actual files. Had to add functions for identifying/reading
whitespace as well, because two consecutive reads will otherwise eat
it up.

I don't have a clear idea in my head on how I want the preprocessing
to happen, and what the output should be (file? string? stream? something else?)
so it's hard to write something I would be satisfied with. I will need to think.