Change the way sexpcodes/forms are parsed/joined
I gave up on trying to optimize too much, and decided to just
CONCATENATE where needed. This actually helped the generated code
become smaller and cleaner, which is nice. I'm not sure about the
performance, but either way I'm not at the stage where that's much of
a concern, since I mostly want to just get something working.

I also made it so that lisp forms need to return a string, a
character, or a number for the result to be embedded. This helps in
avoiding big SYMBOL prints from things like DEFPARAMETER being
embedded in text for example. I might add more allowed types in the
future, but for now this seems good enough.

I'm thinking that building the parse tree might be able to proceed in
tandem with processing, and that higher level constructs like blocks
should be handled in the parser, since that's where control of the
whitespace will be (read at best just stops at it, at worst discards
it, therefore it needs to be wrangled somewhere higher up.) So there
will be no separate processing step that outputs the entire joined
text with values embedded in it, instead every full syntax element
will be tokenized on the spot.