Add functions for peeking/reading characters
Just to save some typing, especially since peek-char has an additional
argument of SKIP-WHITESPACE that I always forget about. As long as
inlining is supported, it shouldn't cause a noticeable impact on
performance in the compiled program, but there is an overhead of
additional function call otherwise. Since the program will eventually
be compiled anyway, it shouldn't be a problem.