Move the constants that declare syntax to their own file
Some of these constants are used in more than one file, so it doesn't
feel right to have it in one or the other, especially if it would
create unwanted dependency on what has to be loaded first. So I moved
all of them to a separate file, which can be loaded first without
problems.
I want to re-use some constants in two different files, so I moved them to a separate one, s