Add escaping of HTML & other minor changes
Wrote a new function to use for reading instead of pulling substrings
out of the buffer; this new function has capabilities to escape
certain characters using html entities, skip over characters that
should not be a part of the html element, etc. Some new constants were
added to accomodate this functionality, and I took the opportunity to
rename some things as well.